% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_operation.R
\name{default_edge_color_aes}
\alias{default_edge_color_aes}
\title{default_edge_color_aes}
\usage{
default_edge_color_aes(g)
}
\arguments{
\item{g}{the graph concerned}
}
\description{
default_edge_color_aes applies a default edge color aesthetics for the graph
}
\examples{
\dontrun{
    default_edge_color_aes(graph)
}
}
