% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_operation.R
\name{default_edge_width_aes}
\alias{default_edge_width_aes}
\title{default_edge_width_aes}
\usage{
default_edge_width_aes(g)
}
\arguments{
\item{g}{the graph concerned}
}
\value{
igraph::graph. The the graph with the ethic.
}
\description{
default_edge_width_aes applies a default edge width aesthetics for the graph
}
\examples{
\dontrun{
    default_edge_width_aes(graph)
}
}
