% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_operation.R
\name{default_selNnSel_vertex_aes}
\alias{default_selNnSel_vertex_aes}
\title{default_selNnSel_vertex_aes}
\usage{
default_selNnSel_vertex_aes(g)
}
\arguments{
\item{g}{the graph concerned}
}
\description{
default_selNnSel_vertex_aes applies a default Sel and NnSel aesthetics for
the vertices of the graph
}
\examples{
\dontrun{
    default_selNnSel_vertex_aes(graph)
}
}
