% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_operation.R
\name{default_vertex_aes}
\alias{default_vertex_aes}
\title{default_vertex_aes}
\usage{
default_vertex_aes(g)
}
\arguments{
\item{g}{the graph concerned}
}
\value{
igraph::graph. The the graph with the ethic.
}
\description{
default_vertex_aes applies a default vertex aesthetics for
the graph
}
\examples{
\dontrun{
    default_vertex_aes(graph)
}
}
