% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-experimental_data.df.r
\docType{data}
\name{experimental_data.df}
\alias{experimental_data.df}
\title{Gérard Derosière's experimental data.}
\format{
A Data frame of Gérard Derosière's experimental data.
}
\source{
Gérard Derosière's experimental data.
}
\usage{
experimental_data.df
}
\description{
It goes without saying that I put them in the package for practical reasons but that they will be removed from the package if the package would be shared if Gérard wishes it. On the other hand if it is not so important considering the evolution of research, the dataframe could be used as an example for the use of the package.
}
\keyword{datasets}
