% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_operation.R
\name{extract_prefix}
\alias{extract_prefix}
\title{extract_prefix}
\usage{
extract_prefix(word, separator)
}
\arguments{
\item{word}{string, the word from which the prefix is to be extracted.}

\item{separator}{string, the separator between the prefix and the word.}
}
\value{
returns the word prefix.
}
\description{
Extract and return the word prefix based on a string separator
}
\details{
extract_prefix split a word an returns the word prefix based on a string.
}
\examples{
str <- "abc"
extract_prefix(str, "b")

str <- "B_column"
extract_prefix(str, "_")
}
