% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_operation.R
\name{graph_structure_construction}
\alias{graph_structure_construction}
\title{graph_structure_construction}
\usage{
graph_structure_construction(dataframe_unit)
}
\arguments{
\item{dataframe_unit}{data.frame. The dataframe unit built using data
separation and prefix functions}
}
\value{
igraph::graph. A graph structure
}
\description{
Graph structure construction returns an igraph::graph structure with all
the attributes of the dataframe unit
}
\details{
graph_structure_construction returns an igraph::graph structure with all
the attributes of the dataframe unit
}
\examples{
\dontrun{
    graph_structure_construction(dataframe_unit)
}
}
