% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_graph_construction.R
\name{grid_graph_automatic_construction}
\alias{grid_graph_automatic_construction}
\title{grid_graph_automatic_construction}
\usage{
grid_graph_automatic_construction(df)
}
\arguments{
\item{df}{the dataframe concerned.}
}
\description{
grid_graph_automatic_construction provides a quick overview of each of
the graphs that can be represented using the dataframe.
}
\details{
Beware the graph ethic depends strongly on the attributes that have been added to your dataframe.
So it goes without saying that this part will certainly have to be customized according to your needs.
This is why there is a detailed guide for more details.
As well as a design thumbnail if you want to know the construction logic behind this package.
}
\examples{
df <- somar::experimental_data.df

grid_graph_automatic_construction(df)
}
