% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_operation.R
\name{layout_in_semicircle}
\alias{layout_in_semicircle}
\title{layout_in_semicircle}
\usage{
layout_in_semicircle(graph)
}
\arguments{
\item{graph}{the input graph.}
}
\value{
A numeric matrix with two columns, and one row for each vertex.
}
\description{
Graph layout with vertices on a semicircle.
}
\details{
Layout in semicircle places vertices on a semicircle, in the order of their vertex ids.
}
\examples{
\dontrun{
    layout_in_semicircle(graph)
}
}
