% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_operation.R
\name{node_vector_from_dataframe_unit}
\alias{node_vector_from_dataframe_unit}
\title{node_vector_from_dataframe_unit}
\usage{
node_vector_from_dataframe_unit(dataframe)
}
\arguments{
\item{dataframe}{data.frame. The dataframe from which the node names must be retrieved}
}
\value{
returns a vector containing the name of all the nodes to be added to the graph.
}
\description{
Allows to retrieve the list of nodes to give to the igraph constructor from a dataframe.
}
\details{
node_vector_from_dataframe_unit returns a vector containing in the order
of the dataframe columns the name of each of the nodes to be displayed on the graph.
}
\examples{
SelIndex_SelThumb <- c(1, 2)
SelIndex_SelPinky <- c(3, 4)

df <- data.frame(SelIndex_SelThumb, SelIndex_SelPinky)
nodes.vec <- node_vector_from_dataframe_unit(df)
}
