% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_operation.R
\name{prefix_vector}
\alias{prefix_vector}
\title{prefix_vector}
\usage{
prefix_vector(dataframe, separator)
}
\arguments{
\item{dataframe}{data.frame. The dataframe from which the prefix vector is to be extracted.}

\item{separator}{string. The prefix separator.}
}
\value{
returns a vector of column prefixes.
}
\description{
Extracts and returns a vector of column prefixes of a dataframe according to a separator
}
\details{
prefix_vector returns a vector of column prefixes of a dataframe according to a separator
}
\examples{
A_column <- c(1, 2, 3)
B_column <- c(4, 5, 6)
C_column <- c(7, 8, 9)
df <- data.frame(A_column, B_column, C_column)

pref_vec <- prefix_vector(df, "_")
}
