% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_data.R
\name{separate_data}
\alias{separate_data}
\title{separate_data}
\usage{
separate_data(dataframe, column_regex_pattern, inverse = FALSE)
}
\arguments{
\item{dataframe}{a dataframe.}

\item{column_regex_pattern}{character string containing a regular expression that represente the column selection.}

\item{inverse}{logical. The function returns the inverse of the result if this parameter is set to true.}
}
\value{
returns a part of the original dataframe
}
\description{
Splits a dataframe in half and returns a new dataframe whose column names
match a certain naming pattern.
}
\details{
separate_data splits a dataframe  in half according to the names of the columns.
This function returns a new dataframe whose column names match a certain naming pattern.
Its behaviour can also be reversed. It is based on column selection using the grep function.
See documentation grep function for more information about regex.
}
\examples{
A_column <- c(1, 2, 3)
B_column <- c(4, 5, 6)
C_column <- c(7, 8, 9)
df <- data.frame(A_column, B_column, C_column)

Bs <- separate_data(df, "^B")
notBs <- separate_data(df, "^B", inverse = TRUE)
}
