% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_operation.R
\name{sequence_order_nodes_according_somatotpy}
\alias{sequence_order_nodes_according_somatotpy}
\title{sequence_order_nodes_according_somatotpy}
\usage{
sequence_order_nodes_according_somatotpy(names)
}
\arguments{
\item{names}{vector. string. the names vector}
}
\value{
returns a position sequence in order to reorder the nodes correctly.
}
\description{
Orders the nodes to reflect the real position of the areas in somatotopy.
}
\details{
The use of this function is specific to a precise naming convention
(see vignette guide) and to the field of somatotopy.
sequence_order_nodes_according_somatotpy returns an order sequence of
the nodes (name vector entered as a parameter) to reflect the real position
of the zones in somatotopy.
}
\examples{
\dontrun{
    sequence_order_nodes_according_somatotpy(columns names)
}
}
