% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-somatotopy.fac.r
\docType{data}
\name{somatotopy.fac}
\alias{somatotopy.fac}
\title{The body area corresponds to a point on the primary somatosensory cortex.}
\format{
A factor of all area.
}
\source{
Wikipedia.
}
\usage{
somatotopy.fac
}
\description{
The body area corresponds to a point on the left primary somatosensory cortex starting from the bottom left in a classical pattern.
}
\keyword{datasets}
