% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-somatotopy.vec.r
\docType{data}
\name{somatotopy.vec}
\alias{somatotopy.vec}
\title{The body area corresponds to a point on the primary somatosensory cortex.}
\format{
A vector of all area.
}
\source{
Wikipedia.
}
\usage{
somatotopy.vec
}
\description{
The body area corresponds to a point on the left primary somatosensory cortex starting from the bottom left in a classical pattern.
}
\keyword{datasets}
