% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_operation.R
\name{substract_prefix}
\alias{substract_prefix}
\title{substract_prefix}
\usage{
substract_prefix(word, prefix)
}
\arguments{
\item{word}{string. The word from which the prefix must be removed.}

\item{prefix}{string. The prefix to be removed from the word.}
}
\value{
returns the word without the prefix.
}
\description{
Subtracts the prefix of the word entered as a parameter
}
\details{
subtracts the prefix of the word entered as a parameter and returns the same word without prefix.
}
\examples{
str <- "abcd"
word <- substract_prefix(str, "ab")

column_name <- "R_column"
word <- substract_prefix(column_name, "R_")
}
