% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_operation.R
\name{substract_prefix_from_all_column}
\alias{substract_prefix_from_all_column}
\title{substract_prefix_from_all_column}
\usage{
substract_prefix_from_all_column(dataframe, prefix)
}
\arguments{
\item{dataframe}{data.frame. The dataframe from which the column prefixes must be removed.}

\item{prefix}{string. The prefix to subtract from column names.}
}
\value{
returns a new dataframe identical but without prefix to the column names.
}
\description{
Subtracts the indicated prefix from all the columns of the dataframe.
}
\details{
substract_prefix_from_all_column subtracts the specified prefix from all the
columns of the dataframe and returns a new dataframe.
}
\examples{
A_Column1 <- c(1, 2)
A_Column2 <- c(3, 4)
A_Column3 <- c(5, 6)
df <- data.frame(A_Column1, A_Column2, A_Column3)

newdf <- substract_prefix_from_all_column(df, "A_")
}
