% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_operation.R
\name{default_graph_aesthetic}
\alias{default_graph_aesthetic}
\title{default_graph_aesthetic}
\usage{
default_graph_aesthetic(g)
}
\arguments{
\item{g}{the graph concerned}
}
\description{
default_graph_aesthetic applies a default aesthetics for the graph
}
\examples{
\dontrun{
    default_graph_aesthetic(graph)
}
}
